add explicit checking of the request method

+++ cinema/views.py
@@ -9,11 +9,11 @@ from cinema.serializes import MovieSerializer
 
 @api_view(["GET", "POST"])
 def movie_list(request):
-    if request.method == "GET":
+    if request.method == "POST":
         movies = Movie.objects.all()
         serializer = MovieSerializer(movies, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
-    else:
+    elif request.method == "GET":
         serializer = MovieSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -31,6 +31,6 @@ def movie_detail(request, pk):
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_200_OK)
-    else:
+    elif request.method == "DELETE":
         movie.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
